% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-mv.R
\name{densidade}
\alias{densidade}
\title{Densidade da normal multivariada}
\usage{
densidade(objeto, y)
}
\arguments{
\item{objeto}{Objeto \code{normal_mv}.}

\item{y}{Ponto (vetor) onde avaliar a densidade.}
}
\value{
Valor da densidade em \code{y}.
}
\description{
Densidade da normal multivariada
}
