% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencia.R
\name{forma_quadratica}
\alias{forma_quadratica}
\title{Forma quadratica y'Ay}
\usage{
forma_quadratica(y, A)
}
\arguments{
\item{y}{Vetor.}

\item{A}{Matriz simetrica.}
}
\value{
Escalar \verb{y'Ay}.
}
\description{
Forma quadratica y'Ay
}
\examples{
forma_quadratica(c(1, 2, 3), diag(3))   # = 14
}
