% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{inversa_minimos_quadrados}
\alias{inversa_minimos_quadrados}
\title{Inversa generalizada de minimos quadrados}
\usage{
inversa_minimos_quadrados(A, explicar = FALSE)
}
\arguments{
\item{A}{Matriz numerica.}

\item{explicar}{Se \code{TRUE}, imprime a verificacao e devolve \code{G} invisivelmente.}
}
\value{
Matriz \code{G} (g-inversa de minimos quadrados de \code{A}).
}
\description{
G-inversa \code{G} tal que \code{A \%*\% G \%*\% A == A} e \code{A \%*\% G} e simetrica. Construida
como \verb{(A'A)^- A'}, usando a g-inversa condicional de \verb{A'A}.
}
