% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnlm-package.R
\docType{package}
\name{learnlm-package}
\alias{learnlm}
\alias{learnlm-package}
\title{learnlm: aprendizado interativo de modelos lineares de posto incompleto}
\description{
Constroi a matriz de delineamento completa, resolve o sistema de equacoes
normais com inversas generalizadas, obtem solucoes sob diferentes restricoes,
testa estimabilidade e narra cada etapa passo a passo.
}
\author{
\strong{Maintainer}: Pedro Mambelli \email{pedromambelli@gmail.com}

Authors:
\itemize{
  \item Pedro Mambelli \email{pedromambelli@gmail.com}
  \item Joao Paulo Bonifacio \email{joao.bonifacio2@estudante.ufla.br}
}

}
\keyword{internal}
