% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-mv.R
\name{marginal}
\alias{marginal}
\title{Marginal de uma normal multivariada}
\usage{
marginal(objeto, indices)
}
\arguments{
\item{objeto}{Objeto \code{normal_mv}.}

\item{indices}{Indices das componentes a manter.}
}
\value{
Novo objeto \code{normal_mv}.
}
\description{
Marginal de uma normal multivariada
}
