% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teoremas.R
\name{numero_funcoes_estimaveis}
\alias{numero_funcoes_estimaveis}
\title{Numero maximo de funcoes estimaveis linearmente independentes}
\usage{
numero_funcoes_estimaveis(objeto)
}
\arguments{
\item{objeto}{Objeto \code{llm}.}
}
\value{
Inteiro \code{posto(X)}.
}
\description{
Pelo teorema de estimabilidade, esse numero e igual a \code{posto(X)}.
}
