% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{plot.regiao_llm}
\alias{plot.regiao_llm}
\title{Grafico da regiao de confianca (2 funcoes)}
\usage{
\method{plot}{regiao_llm}(x, rotulos = c("psi1", "psi2"), ...)
}
\arguments{
\item{x}{Objeto \code{regiao_llm} (de \code{\link[=regiao_confianca]{regiao_confianca()}}).}

\item{rotulos}{Rotulos dos eixos (padrao \code{c("psi1", "psi2")}).}

\item{...}{Passado a \code{plot}.}
}
\value{
Invisivelmente, o objeto.
}
\description{
Desenha a elipse de confianca quando ha exatamente duas funcoes estimaveis.
}
