% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{posto}
\alias{posto}
\title{Posto de uma matriz}
\usage{
posto(A, tol = 1e-07)
}
\arguments{
\item{A}{Matriz numerica.}

\item{tol}{Tolerancia para o posto.}
}
\value{
Inteiro com o posto de \code{A}.
}
\description{
Posto de uma matriz
}
\examples{
posto(matrix(c(1, 2, 2, 4), 2))
}
